InsightApps/requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
# boto3==1.34.99
# botocore==1.34.99
boto3
botocore
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
cx-Oracle==8.3.0
Django==4.1.13
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-rest-authtoken==2.1.4
django-rest-knox==4.2.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.15.4
greenlet==3.0.3
ibm-db-sa==0.4.0
ibm_db==3.2.3
idna==3.7
inflection==0.5.1
isodate==0.6.1
jmespath==1.0.1
jwcrypto==1.5.6
lxml==5.2.2
more-itertools==10.3.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
odbcinst==1.0.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
pkce==1.0.3
platformdirs==4.2.2
plotly==5.22.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pymongo==3.12.3
PyMySQL==1.1.1
pyodbc==5.1.0
pyOpenSSL==24.2.1
PyPDF2==3.0.1
pypdfium2==4.29.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.10.1
simple-salesforce==1.12.6
six==1.16.0
snowflake-connector-python==3.12.0
snowflake-sqlalchemy==1.6.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
sqlalchemy-sybase==2.0.0
sqlglot==25.6.1
sqlparse==0.2.4
tenacity==8.2.3
tomlkit==0.13.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
# urllib3==2.2.1
urllib3
xlrd==2.0.1
zeep==4.2.1
